+++ src/components/input/HelperText.svelte
@@ -1,0 +1,5 @@
+<script>
+	export let helper;
+</script>
+
+{helper}
