+++ src/components/input/InputLine.svelte
@@ -1,0 +1,1 @@
+<input {type} on:focus={focus} on:blur={blur}/>
