+++ src/components/input/README.md
@@ -33,6 +33,7 @@ An input/textarea component with an animated bottom border.
 | `value`                                     | `string`  | `null`                           | The current value of the input.                                                                                                                                                                                         |
 | `placeholder`                               | `string`  | `null`                           | The placeholder and label of the input.                                                                                                                                                                                 |
 | `helper`                                    | `string`  | `null`                           | If specified, UIInput will display a helper text under the input.                                                                                                                                                       |
+| `persistHelper`                             | `boolean` | `false`                          | If `true` UIInput will always show the helper even if the input is not is focus. Only valid if `helper != null`.                                                                                                        |
 | `type`                                      | `string`  | `text`                           | The type of the input.                                                                                                                                                                                                  |
 | `rows`                                      | `number`  | `1`                              | The number of rows in the input.                                                                                                                                                                                        |
 | `compress`                                  | `boolean` | `true`                           | If the input should have no more lines than needed. Only valid if `rows >= 2`.                                                                                                                                          |
